[INLONG-9402][Sort] Add version class to compatible.
